requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
torch>=2.0.0
torchvision>=0.15.0
pandas>=1.2.0
scikit-learn>=0.24.0
Pillow>=10.2.0
matplotlib>=3.7.2
seaborn>=0.12.2
requests>=2.31.0
imutils
deepface 